Feat(modularize): Modularize TodoForm.
